backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.4.0
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.3.9
pydantic==2.5.0